requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
app==0.0.1
astroid==3.3.5
async-timeout==5.0.1
asyncpg==0.30.0
black==24.10.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
databases==0.9.0
dill==0.3.9
fastapi==0.115.4
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
pydantic==2.9.2
pydantic_core==2.23.4
pylint==3.3.1
pytest==8.3.3
pytest-asyncio==0.24.0
python-dotenv==1.0.1
redis==5.0.0
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.2
tomlkit==0.13.2
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
